Add files via upload
I've designed this project because of the security of wild animals. By prevention of bush burner that can harm those animals this system has the solution to prevent them by detecting in case they try to cross the limits of park system will notify the Gard on their screen that there is someone trying to cross the limits and verse vase to animals.